+++ .eslintrc.cjs
@@ -47,5 +47,5 @@ module.exports = {
         ],
     },
 
-    ignorePatterns: ['.eslintrc.cjs'],
+    ignorePatterns: ['.eslintrc.cjs', 'prettier.config.js'],
 };
