+++ package.json
@@ -12,8 +12,9 @@
         "test": "jest"
     },
     "dependencies": {
-        "@aws-sdk/client-s3": "^3.540.0",
+        "@aws-sdk/client-s3": "^3.550.0",
         "@fastify/cors": "^9.0.1",
+        "@sindresorhus/fnv1a": "^3.1.0",
         "@turf/meta": "^6.5.0",
         "dotenv": "^16.4.5",
         "fastify": "^4.26.2",
@@ -21,7 +22,7 @@
     },
     "devDependencies": {
         "@types/geojson": "^7946.0.14",
-        "@types/node": "^20.12.3",
+        "@types/node": "^20.12.5",
         "@typescript-eslint/eslint-plugin": "^7.5.0",
         "@typescript-eslint/parser": "^7.5.0",
         "concurrently": "^8.2.2",
@@ -32,6 +33,6 @@
         "jest": "^29.7.0",
         "nodemon": "^3.1.0",
         "prettier": "^3.2.5",
-        "typescript": "^5.4.3"
+        "typescript": "^5.4.4"
     }
 }
