--- src/GaiaTileSet/index.ts
@@ -1,1 +1,0 @@
-export {GaiaTileSet} from './GaiaTileSet.js';
