--- src/HGT/index.ts
@@ -1,3 +1,0 @@
-export {fetchHGTData} from './fetchHGTData.js';
-export {getHGTElevation} from './getHGTElevation.js';
-export type {HGTData} from './types.js';
